optimize logic
